# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BotAiMode } from "./botAiMode";
import type { BotStatus } from "./botStatus";

export interface Bot {
  id: number;
  name: string;
  serverHost: string;
  serverPort: number;
  version: string;
  status: BotStatus;
  /** @nullable */
  hp?: number | null;
  /** @nullable */
  hunger?: number | null;
  /** @nullable */
  xp?: number | null;
  /** @nullable */
  armor?: number | null;
  /** @nullable */
  posX?: number | null;
  /** @nullable */
  posY?: number | null;
  /** @nullable */
  posZ?: number | null;
  /** @nullable */
  biome?: string | null;
  /** @nullable */
  inventory?: string | null;
  aiMode: BotAiMode;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  autoReply: boolean;
  autoLogin: boolean;
  /** @nullable */
  proxyHost?: string | null;
  /** @nullable */
  proxyPort?: number | null;
  /** @nullable */
  proxyType?: string | null;
  /** @nullable */
  proxyUser?: string | null;
  useAi: boolean;
  createdAt: string;
  updatedAt: string;
}