# lib/api-zod/src/generated/types/botActionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BotActionBodyAction } from "./botActionBodyAction";

export interface BotActionBody {
  action: BotActionBodyAction;
  /** @nullable */
  value?: string | null;
}