# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageType } from "./chatMessageType";

export interface ChatMessage {
  id: number;
  botId: number;
  content: string;
  sender: string;
  type: ChatMessageType;
  createdAt: string;
}