# lib/api-zod/src/generated/types/createBotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BotAiMode } from "./botAiMode";

export interface CreateBotBody {
  name: string;
  serverHost: string;
  serverPort: number;
  version?: string;
  aiMode?: BotAiMode;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  autoReply?: boolean;
  autoLogin?: boolean;
  /** @nullable */
  proxyHost?: string | null;
  /** @nullable */
  proxyPort?: number | null;
  /** @nullable */
  proxyType?: string | null;
  /** @nullable */
  proxyUser?: string | null;
  useAi?: boolean;
}