# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalBots: number;
  onlineBots: number;
  autonomousBots: number;
  botsWithAi: number;
}